Move payment related code from `PointOfSaleDashboardViewModel` to `TotalsViewModel`.
